#2994 - T4A - Upload File to Student Account - Part 2 (#5511)
- Expose new env variables.
- `T4A_FOLDER`, as planned, was removed from the constant and converted
to an env config.
- `T4A_ARCHIVE_FOLDER` was exposed to allow having all the archived
files added to the same shared folder. Currently, all archive operations
implicitly look for an "Archive" folder at the same level as the files
being processed, which would not work for the T4A and its subfolder,
unless an agreement is made to create them ahead of time (either
manually or automatically), which is not desired.
- File group adjusted to "Tax Documents".
- Introduced a small helper to allow some level of identification of SSH
issues. The initial goal is to detect a permanent issue where the
processing of a T4A batch should be stopped because no further attempt
will likely work.
- Method `renameFile` from the SSH changed to allow using an existing
SFTP client.
- Method `archiveFile` changed to allow it to optionally receive the
"absolute" path where the file should be archived.
- Minor and basic SIN validation added to the SIN number received to
ensure they are in the expected format, since the job can receive manual
input.

## E2E Tests
Queue processor for t4a-upload-enqueuer,
√ Should enqueue 3 batches when the batch size is 100, and there are 2
folders, one with 100 files and the other with 101 files.
√ Should log that no directories were found when no directories were
listed from the SFTP.
√ Should not try to add bulk jobs when a directory was found from the
SFTP, but no files were found.

Queue processor for t4a-upload,
√ Should download a T4A file from the SFTP, create a file in the student
account, upload the file to S3 storage, queue a virus scan, send a
notification to the student, and archive the processed file on success
when the student has the current SIN valid, and matching the file name.
√ Should find a student match when the SIN is valid for one of the SIN
validations entries when the student has multiple valid SIN numbers.
√ Should skip the file upload when a file with the same hash was already
uploaded for the same student.
√ Should log a warning and archive the file with a not found student
when an invalid SIN is found in the T4A file name.
√ Should log a warning and archive the file when multiple student
accounts have the same valid SIN.